filesystem: Trigger Prometheus reload if files changed after timeout
The timeout is pretty important as reloading Prometheus is rather expensive. It may happen that Pyrra generates a couple of files in a short period of time (like at startup) and we want to make sure to wait for a brief moment until we're mostly done and only then reload Prometheus.