Don't round objective status query time
This one I was chasing for weeks now. It turns out that Pyrra often show's "No Data" because the timestamp send to Prometheus is actually too far in the future (at 15:01 it rounds to 15:05) and Prometheus doesn't yet have data.
This was an optimization for when the availability & error budget weren't recording rules yet.
Turns out 5min was way too much anyway...

Hopefully the "no data" is mostly a thing of the past now.