slo: Create absent rules for SLI metrics
If a metric for a SLO and SLI isn't available, called absent, we will fire an alert pretty quickly.
It also should be of severity critical as we aren't able to tell if the error budget is burning too quickly. Hence, we'll treat it like the worst outage for the service possible.