ui/aligneddata: support for no label and multiple labels
Support for graph data that not only has multiple or no labels.

Previously, the graph would not display any data that had no
labels, and in cases where the data had multiple labels, there was a
potential for the labels to be incorrectly attached to wrong time series
in the graph. However, it is unclear if the multi-labels scenario
actually occurs in practice.

Signed-off-by: Julien Pivotto <roidelapluie@o11y.eu>