This issue was uncovered after activating the generation of generic rules, one error budget in particular was extremely low on Grafana but all OK according to pyrra / manual checks.
We have 2 SLOs setup for one of our application:

```yaml
  indicator:
    ratio:
      errors:
        metric: grpc_server_processing_duration_seconds_count{app_kubernetes_io_name="grpc_server", statusCode!="OK"}
      total:
        metric: grpc_server_processing_duration_seconds_count{app_kubernetes_io_name="grpc_server"}
  target: "99"
  window: 2w
```

And a second one:

```yaml
  indicator:
    latency:
      success:
        metric: grpc_server_processing_duration_seconds_bucket{app_kubernetes_io_name="grpc_server",
          le="0.1"}
      total:
        metric: grpc_server_processing_duration_seconds_count{app_kubernetes_io_name="grpc_server"}
  target: "95"
  window: 2w
```

Both rules create recording rules named `grpc_server_processing_duration_seconds:increase2w`, as expected. The only way to differentiate the values to use for a specific SLO is to match on the SLO name.

The generic rule for latency was working perfectly, but not the ratio one: the matcher on SLO was missing.

This is what this PR does: add the matching on SLO name for generic rules generated from ratio SLOs.