Improve SLOMetricAbsent alert wait duration based on objective
Previously, Pyrra only took the objective's time window into account. However, there's a huge difference when it comes to burning 1% of error budget for a SLO with 90%,99%,99.5% objective.

The new AbsentDuration function takes both the window and objective's target into account and based on that returns the duration in which 1% error budget would be burned, assuming that metrics absent is a 100% error rate.