Generic Rules for Bool Gauge SLOs
When trying to define a bool_gauge SLO with generic rules enabled, the resulting prometheus rule CRD is invalid as the generic rule group is empty.

This adds support for generic rules for bool gauges SLOs.

It also changes the definition of bool_gauge SLO a bit, by making grouping optional (in line with the rest of the SLOs definitions).