fix: stringify pyrra_window expression
While trying to use [Azure Prometheus Rules Converter](https://github.com/Azure/prometheus-collector/tree/main/tools/az-prom-rules-converter#az-prom-rules-converter) with generic rules enabled, the following error was thrown:

```
Failed to validate Prometheus Rules Schema
[
  {
    "instancePath": "/groups/2/rules/1/expr",
    "schemaPath": "#/properties/expr/type",
    "keyword": "type",
    "params": {
      "type": "string"
    },
    "message": "must be string"
  },
]
```

This corresponds to the `pyrra_window` expression, which is currently a raw integer. Turning it into a string fixes the issue and does not seem to have any impact on the rule on Prometheus. This is also in line with the fact that `pyrra_objective` is also stringified.