Improve Absent alert duration
This PR introduces the AbsentDuration method.
AbsentDuration calculates the duration when absent alerts should fire. The idea is as follows:
Use the most critical of the multi burn rate alerts. For that alert to fire, both the short AND long windows have to be above the threshold. The long window takes the - longest - to fire. Assuming absence of the metric means 100% error rate, the time it takes to fire is the duration for the long window to go above the threshold (factor * objective). Finally, we add the "for" duration we add to the multi burn rate alerts.

Taking the windows and the objective into account the absent alert depends on your SLO's objective and window.

Looking at the tests the duration significantly increased for most absent alerts.