fix(main): Fix nil pointer reference in GetStatus()
We explicitly check for existence in status before adding this metric,
but then we reference the status we know we didn't retrieve. I think
the broader bug might be that the fingerprints for the total value is
different than the one for the errors, so we might need to collate
by closest objective label or something, but this should fix the Nil
pointer exceptions.

fixes #1393