Support empty histogram series
Occasionally a histogram can be created with no data in (for example, an application might have a 5xx series with no data in, so that increases can be caluclated)
Using `histogram_fraction` against series that contains NaN results in NaN, even if other series exist. However, if we sum before histogram_fraction, then the NaNs are essentially ignored.